Offline filtering: split
modified offline filtering to be separated for quirks and events
removed the filterval field from QuirksActivity because it is only by all or today
the filtered lists are not applied to the listview yet
still have to implement the today and comment filters
